Pret avec tout les conditions necessais

+++ Bibliotheque/src/main/java/com/bibliotheque/Bibliotheque/repository/ProfilRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bibliotheque.Bibliotheque.model.Profil;
 
-public interface ProfilRepository extends JpaRepository<Profil, Long> {
+public interface ProfilRepository extends JpaRepository<Profil, Integer> {
 }
